It's commit time, boys and girls.
The GSSM system now has 25 activated map calculations in both
gssm-interface and gssm-autopilot.  A lot of memoization has been
built both into the simulation code and the analyzer functions.

However, the actual and blocked maps do not currently produce useful
output, and although they no longer blow the stack, they will hang
forever in a pointless infinite loop.  That is, DON'T run them yet.

This is a commit for safety purposes, to record the current stable
state of the model in the event of any local damage to the codebase.

The bad news is that after many hours of hacking on the heavily
mutually recursive function that all the actual and blocked map
calculations depend on, I've managed to exhaust all of the obvious
programming tricks that might seem to relieve the imminent stack
blowing of doom (converting to self-recursion, trampolines,
memoization, and using lazy sequences).

The good news is that, as I mentioned above, we're not blowing stack
anymore.  Whoohoo! :) This is because I've had to invent entirely new
and frightening ways to complete this calculation using combinations
of the the above recursive techniques to make the magic move forward.

We're getting pretty close to a final solution at this point, but
after 14 straight hours of working on this problem, my brain has stop
spitting out useful information.

Until I resume again, good night and good luck.


git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/aries@2236 ca20354d-2f28-0410-8711-9259e250f5a2