Fixed the mysterious memory leak in the viewshed model.  It was a degenerate call in rv-resample, which was trying to divide up a distribution into partitions of size 0, hence creating an infinite number of partitions.  A very tough to trace bug.
There were also some edge case errors in rv-lt, rv-gt,
rv-scalar-multiply, rv-scalar-divide, and scalar-rv-divide, which have
since been repaired.

Finally, I tuned the line-of-sight-model's decay and undecay functions
to allow paths to travel a bit farther.  This still needs to be
matched to the resolution of the underlying data, but that's a problem
for another day.


git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/aries@2586 ca20354d-2f28-0410-8711-9259e250f5a2