SPAN model runs and returns; result observation generated (no more null pointer exceptions). Still need to insert dependencies based on results - must clear up a few doubts before that, but it should be SOON.
Elegant mechanism to define the observables for the different SPAN results - whatever concept is subsumed by any of the {Source,Sink,Use,Flow}Traits in the concept space of the corresponding observables becomes an observation for the corresponding result; if nothing implements that trait, its observation is not generated. This gives total control to the ontologies over what gets computed or not, and inserts the self-documenting observables. I only set traits in the aestheticService ontology, and not all of them for now.

The span model form requires BOTH a flow-concept (to generate the flow observations) and a flow-data-concept (to provide data to the flow model). span-driver doesn't need flow-concept, only flow-data-concept, so it keeps working as before.

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/aries@2589 ca20354d-2f28-0410-8711-9259e250f5a2