* Changes to clj-misc.matrix-ops
Added a suite of functions for 2D linear geometry:
  get-slope, get-y-intercept, line, get-line,
  get-line-fn, normal-slope, project-onto-line

Began work on find-cells-along-arc
  (for coastal-storm-protection-model).

* Changes to clj-span.carbon-model

Added use-percents and actual-sinks-by-source the carrier-lists
calculation.

Added a combine worlds function to merge multiple MC simulation runs.

Added animation? to the distribute-flow input args to reflect its
updated definition.

Updated distribute flow to return (all zero) possible-flow and
actual-flow layers.

* Changes to clj-span.coastal-storm-protection-model

Imported euclidean-distance, get-line, project-onto-line, and
find-cells-along-arc.

Renamed determine-mean-storm-bearing to sample-storm-track and changed
it to return the storm-track-sample and lead-id (of this sample).

Changed get-bearing-seq to return pairs of cell-id and bearing instead
of simply the bearings as before.

Changed find-wave-cells to return a sequence of pairs [cell-id
distance-from-centerpoint].

Added a :centerpoint-offset field (distance from the
storm-centerpoint) to each storm-carrier.

Renamed apply-local-effects! to explore-next-segment! and began
extending it to allow carriers to explore cell arcs of different
lengths as the wave turns.

Patched get-next-bearing to prevent the algorithm from letting
carriers pivot by a full 180%.

Passed mean-storm-bearing, storm-track-sample, and lead-id to the wave
movement function in distribute-wave-energy!

Factored the iterated movement function out of
distribute-wave-energy!, named it advace-wave, and began extending it
to enable the wave to pivot incrementally around the
storm-centerpoint as it moves.

* Changes to clj-span.repl-utils

Set source/sink/use-thresholds to 0.0 in test-run-storm.

Added test-run-water.

* Changes to clj-span.surface-flow-water-model
Fixed a bug with properly initializing the carriers' bearing values.

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/aries@3656 ca20354d-2f28-0410-8711-9259e250f5a2